Add a travis build of subsurface
This runs a subsurface script/build.sh build in travis-ci, and runs the
tests afterwards.

The build runs on the Ubuntu Trusty image, but due to the fact that the
Qt shipped there is to old, it installs a Qt 5.8 from qt.io , and with
some trickery caches it.

Hacked out are things that doesn't build with Qt 5.8, and the rest is
built against WebEngine.

The tests currently fail, and I really don't know why, but its a clear
indication that they aren't run that often. This cam makes sure they are
run at least. The actual testing is just commented out for that reason.

Signed-off-by: Anton Lundin <glance@acc.umu.se>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>