user manual: do not repeat
Do not repeat the almost same text in the next sentence.

When entering a dive location name, auto location of
dive site names makes it easy to select an existing dive site name.

When typing the name of a dive site,
a dropdown list appears showing all sites with similar names.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>