user manual: info box does not change
The info box for the dive profile is said to change to minimal when the
mouse pointer is outside the profile window. Testing this in 4.6.3 shows
that this is not the case. The info box does not change on window leave
of the pointer.

It might have worked that way before, but it is no big deal, so better to
change the manual to reflect the current working.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>