user manual: TTS does not depend on SAC
This one little sentence confuses me: TTS is calculated assuming an ascent
surface air consumption (SAC) for the gas currently used.

Time to surface is a decompression + ascent time calculation only. Obviously, the diver
should plan to have suffient gas for this, but TTS is not a function of SAC. There is
the "remaining air time, RAT) some dive computers can calculate which uses TTS in
combination with SAC to assist the diver to start his ascent, but RAT != TTS.

Solution: remove the confusing sentence.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>