Add double imperial tanks to defaults
When not using a dive computer that is air integrated, resetting the tank
size and nomenclature on each dive is tedious for imperial double tanks.
This change sets a standard for their nomenclature that works with the
XML export, and adds the typical tanks doubled by imperial divers to the
default list.

Signed-off-by: Mark Fox <mark.p.fox@gmail.com>