user manual: weird and incorrect paragraph
from the planner section the follwing paragraph is weird and partially incorrect:

These changes should reflect the cylinders and gas compositions defined in the table
with _Available Gases_. If two or more gases are used, automatic gas switches will be
suggested during the ascent to the surface. These changes can be deleted by
right-clicking the gas change and manually creating a gas change by right-clicking
on the appropriate waypoint.

Yes: during the ascent, gas switches are automatically planned (not suggested). These
can't be deleted by right-clicking (as there are no real waypoints associated with
the gas switches). And futher. To delete a gas change you have to manually create one?

Solution: refrase paragraph.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>